Tests moved to proper package
